+++ Importify.Service/Model/CommonImport.cs
@@ -9,7 +9,7 @@ namespace Importify.Service.Model
     public class CommonImport
     {
         public int CommonImportId { get; set; }
-        public long Value { get; set; }
+        public decimal Value { get; set; }
         public Country? Country { get; set; }
         public Year? Year { get; set; }
     }
